feat: ajouter des types pour les pilotes, les sessions de course et les résultats de position refactor: mettre à jour les imports pour utiliser les nouveaux types dans les composants de rendu
